Fixed Notification CSS, Remove Friend, Sending Notification
The code checks if they both follow each other, then they are friends. If only request is sent, then the user can accept or decline the request. If the user is removed from friends, notficiaton is sent. If the user first remove the friend and then send a request, latest request is shown in notificaiton from that specific user.